Fix escaping bug.
When we have to delimiting characters, then we escape them by escaping
the first character. For example if our delimiter is ]#, then we
can get an escaped version like this: "foo \\]#]#" and get "foo ]#"
out of it. If, on the other hand we parse "foo ]\\#]#", then we get
"foo ]\\#". To produce "foo \\]#" we need to parse "foo \\\\\\]#"